Implement spea2 class hierarchy
