+++ src/evolution/spea2/spea2.py
@@ -3,8 +3,6 @@ import random
 from math import sqrt
 from operator import itemgetter
 
-import numpy as np
-
 from src.evolution.raw_fitness import raw_fitness
 
 
@@ -13,10 +11,8 @@ class SPEA2:
         '''
          Strength Pareto Evolutionary Algorithm
         :param params: Meta-parameters of the SPEA2
-        :param init_population: function to generate initial population
         :param objectives: function to calculate objective functions for each individual in population
-        :param crossover: function to crossover two genotypes
-        :param mutation: function to mutate genotype
+        :param evolutionary_operators: EvoOperators class that encapsulates all evolutionary operators
         '''
 
         self.params = params
@@ -90,38 +86,7 @@ class SPEA2:
             return SPEA2.ErrorHistory.Point() if len(self.history) == 0 else self.history[-1]
 
     def solution(self, verbose=True, **kwargs):
-        archive_history = []
-        history = SPEA2.ErrorHistory()
-
-        gen = 0
-        while gen < self.params.max_gens:
-            self.fitness()
-            self._archive = self.environmental_selection(self._pop, self._archive)
-            best = sorted(self._archive, key=lambda p: mean_obj(p))[0]
-
-            last_fit = history.last().fitness_value
-            if last_fit > mean_obj(best):
-                best_gens = best.genotype
-
-                if verbose:
-                    if 'print_fun' in kwargs:
-                        kwargs['print_fun'](best, gen)
-                    else:
-                        print_new_best_individ(best, gen)
-
-                history.add_new(best_gens, gen, mean_obj(best),
-                                rmse(best))
-
-            selected = self.selected(self.params.pop_size, self._archive)
-            self._pop = self.reproduce(selected, self.params.pop_size)
-
-            to_add = copy.deepcopy(self._archive + self._pop)
-            self.objectives(to_add)
-            archive_history.append(to_add)
-
-            gen += 1
-
-        return history, archive_history
+        pass
 
     def fitness(self):
         self.objectives(self._pop)
@@ -160,7 +125,6 @@ class SPEA2:
 
     def environmental_selection(self, pop, archive):
         union = archive + pop
-        # TODO: check value of fitness for union
         env = [p for p in union if p.fitness() < 1.0]
 
         if len(env) < self.params.archive_size:
@@ -169,7 +133,6 @@ class SPEA2:
             for p in union:
                 if len(env) >= self.params.archive_size:
                     break
-                # TODO: check value of fitness for union
                 if p.fitness() >= 1.0:
                     env.append(p)
         elif len(env) > self.params.archive_size:
@@ -182,7 +145,6 @@ class SPEA2:
                     for p2 in env:
                         distances_to_p1.append(self.euclidean_distance(p1.objectives, p2.objectives))
                     distances_to_p1 = sorted(distances_to_p1)
-                    # TODO: check density formula
                     density = 1.0 / (distances_to_p1[k] + 2.0)
                     dens.append((p1, density))
                 dens.sort(key=itemgetter(1))
@@ -229,23 +191,3 @@ class SPEA2:
                 break
 
         return children
-
-
-def rmse(individ):
-    result = 0.0
-    for obj in individ.objectives:
-        result += pow(obj, 2)
-    return sqrt(result / len(individ.objectives))
-
-
-def mean_obj(individ):
-    return np.mean(individ.objectives)
-
-
-def print_new_best_individ(best, gen_index):
-    print("new best: ", round(best.fitness(), 5), round(best.genotype.drf, 2),
-          round(best.genotype.cfw, 6),
-          round(best.genotype.stpm, 6),
-          round(rmse(best), 4),
-          mean_obj(best))
-    print(gen_index)
